+++ Enemy.cpp
@@ -43,7 +43,7 @@ void Enemy::SetPosition(const DirectX::XMFLOAT3& newPosition)
 	//int attackDelay = 4;		//How long between each attack
 	
 	// ATTACK DELAY & DURATION HELD IN CONSTANTS
-	times.StartTime(GC::ENEMYTYPE1_ATTACK_DURATION, GC::ENEMYTYPE1_ATTACK_DELAY);
+	times.StartTime(static_cast<int>(GC::ENEMYTYPE1_ATTACK_DURATION), static_cast<int>(GC::ENEMYTYPE1_ATTACK_DELAY));
 
 	//Setup the enemy particles
 	for (int i = 1; i != 20; i++)
